algunas mejroas

+++ arch_install.sh
@@ -1,46 +1,33 @@
 #!/bin/sh
 
+set -e # Abortará el script si algún comando falla
+
 echo "--------------------------------------------------------------------------------------------------------"
 echo "Instalation of Arch Linux"
 echo "--------------------------------------------------------------------------------------------------------"
 
-# Config
+# Configuración
 TARGET_DISK='/dev/sda'
-# FONT='ter-132b'
-TIMEZONE=$1 
+TIMEZONE=$1
 HOSTNAME=$2
 PASSWD=$3
 PASSWDROOT=$4
 
-if lsblk | grep -q "/mnt"; then
-    umount /mnt/boot
-    umount /mnt
-    swapoff "$TARGET_DISK"2
+if [ -z "$TIMEZONE" ] || [ -z "$HOSTNAME" ] || [ -z "$PASSWD" ] || [ -z "$PASSWDROOT" ]; then
+    echo "Faltan parámetros. Uso: $0 <TIMEZONE> <HOSTNAME> <PASSWD> <PASSWDROOT>"
+    exit 1
 fi
 
-setfont "$FONT"
-timedatectl set-timezone "$TIMEZONE"
-
-# Limpiar disco por si hay algo antes -- Ojito con esto -- 
-sgdisk --zap-all "$TARGET_DISK"
-echo "--------------------------------------------------------------------------------------------------------"
-
-# Disk
+# Limpiar disco
+echo "Limpiando el disco..."
 parted "$TARGET_DISK" mklabel gpt
-echo "--------------------------------------------------------------------------------------------------------"
 
-# Efi system
+
+# Crear particiones
 parted "$TARGET_DISK" mkpart primary fat32 1MiB 501MiB
 parted "$TARGET_DISK" set 1 esp on 
-echo "--------------------------------------------------------------------------------------------------------"
-
-# Swap
 parted "$TARGET_DISK" mkpart primary linux-swap 501MiB 8693MiB
-echo "--------------------------------------------------------------------------------------------------------"
-
-# Linux filesystem
 parted "$TARGET_DISK" mkpart primary ext4 8693MiB 100%
-echo "--------------------------------------------------------------------------------------------------------"
 
 # Formatear particiones
 mkfs.fat -F 32 "$TARGET_DISK"1
@@ -52,31 +39,24 @@ mount "$TARGET_DISK"3 /mnt
 mount --mkdir "$TARGET_DISK"1 /mnt/boot
 swapon "$TARGET_DISK"2 
 
-# Mirrorlist
-pacman -Sy
+# Actualizar mirrorlist
+pacman -Sy --noconfirm
 pacman -S rsync --noconfirm
 reflector --verbose --latest 10 --sort rate --save /etc/pacman.d/mirrorlist
 
 # Instalar componentes basicos
-pacstrap -K /mnt base base-devel linux-zen linux-zen-headers linux-firmware amd-ucode networkmanager xf86-input-libinput mkinitcpio
+pacstrap -K /mnt base base-devel linux-zen linux-zen-headers linux-firmware
 
 # Fstab
 genfstab -p /mnt >> /mnt/etc/fstab
 
 # Entrar al sistema montado
 arch-chroot /mnt <<EOF
-
-# Configurar zona horaria
 ln -sf /usr/share/zoneinfo/${TIMEZONE} /etc/localtime
 hwclock --systohc
 timedatectl set-ntp true
 
-# Instalar y habilitar NTP
-pacman -S --noconfirm ntp
-systemctl enable ntpd
-systemctl start ntpd
-
-# Configurar nombre de host
+# Configuracion de hostname y hosts
 echo "${HOSTNAME}" > /etc/hostname
 
 # Configurar /etc/hosts
@@ -92,7 +72,8 @@ mkinitcpio -P
 # Sudoers
 sed -i '0,/# %wheel/s//%wheel/' /etc/sudoers
 
-# GRUB CONFIGURATION
+# GRUB
+pacman -Syu --noconfirm
 pacman -S grub efibootmgr --noconfirm
 grub-install --target=x86_64-efi --efi-directory=/boot --bootloader-id=Arch --removable
 
@@ -103,10 +84,10 @@ GRUB_HIDDEN_TIMEOUT=0 # Comment if you dont want it\n' /etc/default/grub
 
 grub-mkconfig -o /boot/grub/grub.cfg
 
-# Enable networkmanager
+# Habilitar networkmanager
 systemctl enable NetworkManager
 
-# Pacman configuration
+# Configuration de pacman
 sed -i 's/#Color/Color/' /etc/pacman.conf
 sed -i '/Color/a\ILoveCandy' /etc/pacman.conf
 sed -i 's/#VerbosePkgLists/VerbosePkgLists/' /etc/pacman.conf
@@ -123,10 +104,9 @@ awk '
 { print }
 ' /etc/pacman.conf > tmp && mv tmp /etc/pacman.conf
 
-pacman -S neovim net-tools ly xdg-user-dirs git --noconfirm
+# Instalar paquetes adicionales
+pacman -S neovim net-tools ly xdg-user-dirs git amd-ucode networkmanager xf86-input-libinput mkinitcpio bluez bluez-utils blueman --noconfirm
 systemctl enable ly.service
-
-pacman -S bluez bluez-utils blueman --noconfirm
 systemctl enable bluetooth
 
 pacman -S tlp tlp-rdw powertop acpi --noconfirm
@@ -136,6 +116,10 @@ systemctl mask systemd-rfkill.service
 systemctl mask systemd-rfkill.socket
 
 systemctl enable fstrim.timer
+
+pacman -S ntp --noconfirm 
+systemctl enable ntpd
+systemctl start ntpd
 
 pacman -S xorg-server xorg-apps xorg-xinit --noconfirm
 pacman -S i3-gaps i3blocks i3lock numlockx --noconfirm
@@ -166,12 +150,8 @@ cd Sources
 
 git clone https://aur.archlinux.org/paru.git
 cd paru
-makepkg -si
-
-
-
+makepkg -si --noconfirm
 
 EOF
 
 # # falta instalar bluez bluez-utils umount y pacman config
-
